move from 'js' to 'js_interop' to satisfy conditional_srcs_processor
