+++ packages/flutter/lib/src/painting/_load_asset_manifest_io.dart
