+++ packages/flutter_tools/lib/src/web/file_generators/main_dart.dart
@@ -22,7 +22,7 @@ String generateMainDartFile(String appEntrypoint, {
     '',
     "import 'dart:ui_web' as ui_web;",
     "import 'dart:async';",
-    "import 'dart:js' as js;",
+    "import 'dart:js_interop';",
     '',
     "import '$appEntrypoint' as entrypoint;",
     "import '$pluginRegistrantEntrypoint' as pluginRegistrant;",
@@ -30,10 +30,20 @@ String generateMainDartFile(String appEntrypoint, {
     'typedef _UnaryFunction = dynamic Function(List<String> args);',
     'typedef _NullaryFunction = dynamic Function();',
     '',
+    '@JS("window")',
+    'external _DomWindow get _domWindow;',
+    '@JS()',
+    '@staticInterop',
+    'class _DomWindow {}',
+    '',
+    'extension _DomWindowExtension on _DomWindow {',
+    "@JS('_flutter_assetManifestBytes')",
+    '  external  set _flutter_assetManifestAsByteList(value);',
+    '}',
     'Future<void> main() async {',
     '  await ui_web.bootstrapEngine(',
     '    runApp: () {',
-    '      js.context["_flutter_assetManifestAsByteList"] = [${assetManifestBytes.join(',')}];',
+    '      _domWindow._flutter_assetManifestAsByteList = [${assetManifestBytes.join(',')}];',
     '      if (entrypoint.main is _UnaryFunction) {',
     '        return (entrypoint.main as _UnaryFunction)(<String>[]);',
     '      }',
